# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tài liệu API cho BTL Nhóm 2",
        "description": "Đây là tài liệu API cho ứng dụng điểm danh/quản lý lịch dạy.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/makeupclasses": {
            "get": {
                "tags": [
                    "Makeup Classes"
                ],
                "summary": "Lấy DS Lớp dạy bù (Chưa triển khai)",
                "operationId": "getMakeupClassesList",
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Makeup Classes"
                ],
                "summary": "Gửi yêu cầu dạy bù (Dùng route resource /api/makeupclasses)",
                "description": "Route /api/makeup-classes (custom) cũng trỏ về đây",
                "operationId": "storeMakeupClassResource",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "original_schedule_id",
                                    "new_schedule_date",
                                    "new_session",
                                    "new_room_id"
                                ],
                                "properties": {
                                    "original_schedule_id": {
                                        "description": "ID lịch dạy GỐC (buổi nghỉ)",
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "new_schedule_date": {
                                        "description": "Ngày dạy bù (Y-m-d)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-30"
                                    },
                                    "new_session": {
                                        "description": "Ca/tiết dạy bù",
                                        "type": "string",
                                        "example": "3-4"
                                    },
                                    "new_room_id": {
                                        "description": "ID của phòng dạy bù",
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gửi yêu cầu dạy bù thành công!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập"
                    },
                    "422": {
                        "description": "Lỗi validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/makeup-classes": {
            "post": {
                "tags": [
                    "Makeup Classes"
                ],
                "summary": "Gửi yêu cầu dạy bù (Dùng route /api/makeup-classes)",
                "description": "Route /api/makeupclasses (resource) cũng trỏ về đây",
                "operationId": "storeMakeupClass",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "original_schedule_id",
                                    "new_schedule_date",
                                    "new_session",
                                    "new_room_id"
                                ],
                                "properties": {
                                    "original_schedule_id": {
                                        "description": "ID lịch dạy GỐC (buổi nghỉ)",
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "new_schedule_date": {
                                        "description": "Ngày dạy bù (Y-m-d)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-30"
                                    },
                                    "new_session": {
                                        "description": "Ca/tiết dạy bù",
                                        "type": "string",
                                        "example": "3-4"
                                    },
                                    "new_room_id": {
                                        "description": "ID của phòng dạy bù",
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gửi yêu cầu dạy bù thành công!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập"
                    },
                    "422": {
                        "description": "Lỗi validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/makeupclasses/{makeupclass}": {
            "get": {
                "tags": [
                    "Makeup Classes"
                ],
                "summary": "Xem 1 Lớp bù (Chưa triển khai)",
                "operationId": "getMakeupClassById",
                "parameters": [
                    {
                        "name": "makeupclass",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Makeup Classes"
                ],
                "summary": "Cập nhật Lớp bù (Chưa triển khai)",
                "operationId": "updateMakeupClass",
                "parameters": [
                    {
                        "name": "makeupclass",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Makeup Classes"
                ],
                "summary": "Xóa Lớp bù (Chưa triển khai)",
                "operationId": "deleteMakeupClass",
                "parameters": [
                    {
                        "name": "makeupclass",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sử dụng Bearer token (Sanctum) sau khi đăng nhập",
                "name": "Bearer Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}